OFBIZ-2565: fix for data type mismatch
git-svn-id: https://svn.apache.org/repos/asf/ofbiz/trunk@1214367 13f79535-47bb-0310-9956-ffa450edef68

+++ specialpurpose/hhfacility/src/org/ofbiz/hhfacility/FacilityServices.java
@@ -19,6 +19,7 @@
 
 package org.ofbiz.hhfacility;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
@@ -107,7 +108,7 @@ public class FacilityServices {
         String facilityId = (String) context.get("facilityId");
         String productId = (String) context.get("productId");
         String locationSeqId = (String) context.get("locationSeqId");
-        Double quantity = (Double) context.get("quantity");
+        BigDecimal quantity = (BigDecimal) context.get("quantity");
         if (UtilValidate.isEmpty(productId) || UtilValidate.isEmpty(facilityId)) {
             return ServiceUtil.returnError(UtilProperties.getMessage(resource, "ProductIdOrFacilityIdNotFound", locale));
         }
@@ -127,7 +128,7 @@ public class FacilityServices {
             Debug.logError(e, "updateProductStocktake failed getting inventory counts", module);
             return ServiceUtil.returnError(UtilProperties.getMessage(resource, "ProductErrorUpdateProductStockTake", locale));
         }
-        int qoh = ((Double)invAvailability.get("quantityOnHandTotal")).intValue();
+        int qoh = ((BigDecimal)invAvailability.get("quantityOnHandTotal")).intValue();
         if (quantity.intValue() == qoh) {
             // No change required.
             Debug.logInfo("updateProductStocktake No change required quantity("+quantity+") = qoh("+qoh+")", module);
